requirements-colab.txt
# Colab already ships with torch + CUDA. Just install the rest.
segmentation-models-pytorch>=0.3.3
timm>=0.9.12
albumentations>=1.4.0
torchmetrics>=1.3.0
torch-fidelity>=0.3.0
lpips>=0.1.4
einops>=0.7.0
hydra-core>=1.3.2
huggingface_hub>=0.20.0
